Clean up from test.
The constructor test was not cleaning up. Since there were two test
classes for a single test, combine them into a single test.
